requirements.txt
python-telegram-bot==20.7
qrcode==7.4.2
Pillow==10.1.0 